handle serial disconnects. ignore uninit error.
During some testing, it was found that if a USB TTl cable was
disconnected while in a select() call, it would just return
immediately with 1, indicating that the requested number of bytes
was available. Put in a check for this edge case. I'll have to look
into whether or not there is a smarter way to do this with a system
call. Also removed some old code and rearrageed a few things.